pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syshardn"
version = "0.1.0"
description = "Multi-platform system hardening tool"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]
keywords = ["security", "hardening", "cis", "compliance", "audit"]
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]
reporting = [
    "reportlab>=4.0.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
syshardn = "syshardn.cli:main"

[project.urls]
Homepage = "https://github.com/aswinr24/syshardn"
Documentation = "https://github.com/aswinr24/syshardn/wiki"
Repository = "https://github.com/aswinr24/syshardn"
"Bug Tracker" = "https://github.com/aswinr24/syshardn/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["syshardn*"]

[tool.setuptools.package-data]
syshardn = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
